Add TRD to CTF reader/writer workflows + unit test
Also, move the CTF workflows code from .h to cxx file (leaving only getXXXSpec declaration)
